Added 1 test and 1 more student with WA + checking for time limits

+++ prog/main.cpp
@@ -1,7 +1,8 @@
 #include <iostream>
+#include <unistd.h>
 
 int main() {
-    int a, b;
+    long long a, b;
     std::cin >> a >> b;
     std::cout << a + b << std::endl;
     return 0;
